Worker/requirements.txt
# Core
celery
redis

# Image processing
opencv-python-headless
numpy<2.0
Pillow
scikit-image

# AI / Deep Learning
diffusers==0.31.0
transformers==4.47.1
accelerate
safetensors
scipy
ftfy

# Real-ESRGAN
basicsr==1.4.2
realesrgan==0.3.0

# Generative erase
simple-lama-inpainting

# Background removal
timm
kornia